(PC-16709) refactor(cookies): move usePostCookiesConsent to api folder
