+++ src/features/auth/AuthContext.tsx
@@ -6,8 +6,8 @@ import { api } from 'api/api'
 import { refreshAccessToken } from 'api/apiHelpers'
 import { SigninResponse } from 'api/gen'
 import { useResetContexts } from 'features/auth/useResetContexts'
+import { usePostCookiesConsent } from 'features/cookies/api/usePostCookiesConsent'
 import { useCookies } from 'features/cookies/useCookies'
-import { usePostCookiesConsent } from 'features/cookies/usePostCookiesConsent'
 import { useAppStateChange } from 'libs/appState'
 import { analytics, LoginRoutineMethod } from 'libs/firebase/analytics'
 import { getAccessTokenStatus, getUserIdFromAccesstoken } from 'libs/jwt'
